Lab2-BuggyCode/requirements.txt
apache-airflow==2.8.0
transformers==4.36.0
torch==2.1.0
datasets==2.16.0
pandas==2.1.0
scikit-learn==1.3.0
pylint==3.0.0
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==4.25.0
psycopg2-binary==2.9.9
langchain-openai==0.1.0
langchain-core>=0.1.33,<0.2.0
python-dotenv==1.0.0